requirements.txt
google-auth==2.17.0
google-auth-oauthlib==0.5.2
google-auth-httplib2==0.1.0
google-api-python-client==2.77.0
moviepy==1.0.3
Pillow==9.0.1
oauth2client
facebook_business
dropbox